debug_performance_analysis.py
#!/usr/bin/env python3
"""
🔍 Debug Performance Analysis
Analyze why Enhanced Nanpin isn't beating Buy & Hold despite leverage and strategic timing
"""

import asyncio
import pandas as pd
import numpy as np
import yfinance as yf
from datetime import datetime, timedelta

async def debug_performance_comparison():
    """Debug why our strategy underperforms Buy & Hold"""
    
    print("🔍 DEBUGGING ENHANCED NANPIN PERFORMANCE")
    print("="*60)
    
    # Load BTC data
    start_date = "2020-01-01"
    end_date = "2024-12-31"
    
    btc = yf.Ticker("BTC-USD")
    data = btc.history(start=start_date, end=end_date, interval="1d")
    btc_data = data[['Close']].copy()
    btc_data.index = pd.to_datetime(btc_data.index.date)
    
    # Calculate Buy & Hold
    start_price = btc_data['Close'].iloc[0]
    end_price = btc_data['Close'].iloc[-1]
    
    print(f"📊 BTC PRICE ANALYSIS:")
    print(f"   Start Price (Jan 1, 2020): ${start_price:,.0f}")
    print(f"   End Price (Dec 31, 2024): ${end_price:,.0f}")
    print(f"   Total Price Appreciation: {(end_price/start_price - 1)*100:.1f}%")
    
    years = (datetime(2024, 12, 31) - datetime(2020, 1, 1)).days / 365.25
    buy_hold_annual = (end_price/start_price)**(1/years) - 1
    print(f"   Buy & Hold Annual Return: {buy_hold_annual*100:.1f}%")
    
    # Simulate our strategy behavior
    print(f"\n🔍 NANPIN STRATEGY ANALYSIS:")
    
    # Example: $100k invested
    total_investment_budget = 100000
    base_amount = 100  # $100 base position
    
    # Buy & Hold comparison
    buy_hold_btc = total_investment_budget / start_price
    buy_hold_final_value = buy_hold_btc * end_price
    buy_hold_return = (buy_hold_final_value - total_investment_budget) / total_investment_budget
    
    print(f"   Buy & Hold with ${total_investment_budget:,}:")
    print(f"     BTC acquired: {buy_hold_btc:.6f}")
    print(f"     Final value: ${buy_hold_final_value:,.0f}")
    print(f"     Total return: {buy_hold_return*100:.1f}%")
    
    # Our strategy simulation
    print(f"\n💡 IDENTIFYING THE ISSUES:")
    
    # Issue 1: We don't invest all capital immediately
    print(f"   ISSUE 1: Capital deployment")
    print(f"     Buy & Hold: Invests ${total_investment_budget:,} immediately at ${start_price:,.0f}")
    print(f"     Our Strategy: Only invests when Fibonacci levels hit")
    print(f"     This means we often miss early gains!")
    
    # Issue 2: We're buying throughout the cycle, including at higher prices
    print(f"\n   ISSUE 2: Average purchase price")
    print(f"     Buy & Hold: Always buys at ${start_price:,.0f}")
    print(f"     Our Strategy: Buys at various prices (some high, some low)")
    print(f"     Even with leverage, higher average price can hurt returns")
    
    # Issue 3: We don't account for the fact that we're deploying capital over time
    print(f"\n   ISSUE 3: Time-weighted returns")
    print(f"     Buy & Hold: Full capital working for full 5 years")
    print(f"     Our Strategy: Capital deployed gradually over time")
    print(f"     Later investments have less time to compound")
    
    # Let's simulate what SHOULD happen with perfect execution
    print(f"\n🚀 WHAT SHOULD HAPPEN WITH PERFECT NANPIN:")
    
    # Simulate buying only at major lows with leverage
    major_lows = [
        {"date": "2020-03-13", "price": 3800, "reason": "COVID crash"},
        {"date": "2020-09-24", "price": 10200, "reason": "Post-crash recovery"},
        {"date": "2022-06-18", "price": 17600, "reason": "Bear market low"},
        {"date": "2022-11-21", "price": 15500, "reason": "FTX collapse"},
    ]
    
    perfect_nanpin_btc = 0
    perfect_investment = 0
    investment_per_dip = total_investment_budget / len(major_lows)
    
    for low in major_lows:
        # Use 5x leverage on each major dip
        leverage = 5
        amount_invested = investment_per_dip * leverage
        btc_bought = amount_invested / low["price"]
        perfect_nanpin_btc += btc_bought
        perfect_investment += investment_per_dip  # Only count actual capital, not leveraged amount
        
        print(f"     {low['date']}: Buy ${amount_invested:,.0f} at ${low['price']:,.0f} ({low['reason']})")
        print(f"       BTC acquired: {btc_bought:.6f}")
    
    perfect_final_value = perfect_nanpin_btc * end_price
    perfect_return = (perfect_final_value - perfect_investment) / perfect_investment
    
    print(f"\n   Perfect Nanpin Results:")
    print(f"     Total invested: ${perfect_investment:,.0f}")
    print(f"     Total BTC: {perfect_nanpin_btc:.6f}")
    print(f"     Final value: ${perfect_final_value:,.0f}")
    print(f"     Total return: {perfect_return*100:.1f}%")
    print(f"     Annual return: {(perfect_final_value/perfect_investment)**(1/years)*100-100:.1f}%")
    
    # The real issue diagnosis
    print(f"\n❌ WHY OUR CURRENT STRATEGY UNDERPERFORMS:")
    print(f"   1. We spread buys across too many price points")
    print(f"   2. We don't use enough leverage consistently")
    print(f"   3. We buy during recoveries, not just major crashes")
    print(f"   4. Our position sizing is too conservative")
    print(f"   5. We don't deploy capital fast enough during opportunities")
    
    print(f"\n✅ SOLUTIONS:")
    print(f"   1. Increase minimum leverage to 5-10x")
    print(f"   2. Only buy during extreme fear periods (F&G < 15)")
    print(f"   3. Larger position sizes during major crashes")
    print(f"   4. Skip small dips, focus on 30%+ corrections")
    print(f"   5. Deploy capital more aggressively during crisis")

async def main():
    await debug_performance_comparison()

if __name__ == "__main__":
    asyncio.run(main())