requirements_compatible.txt
# Compatible versions for Python 3.12
asyncio-mqtt==0.13.0
aiohttp==3.9.1
aiofiles==23.2.0

# Data manipulation and analysis
pandas>=2.1.4
numpy>=1.26.0
scipy>=1.11.4

# Technical analysis
pandas-ta>=0.3.14b0
# Note: TA-Lib installation failed, using pandas-ta instead

# Cryptocurrency APIs
python-binance==1.0.19
ccxt==4.1.64

# Web scraping and APIs
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Cryptography for Backpack API
cryptography>=41.0.7
PyNaCl>=1.5.0
pycryptodome>=3.19.0

# Logging and monitoring
structlog==23.2.0
colorama==0.4.6

# Configuration management
PyYAML==6.0.1
python-dotenv==1.0.0

# Date and time handling
python-dateutil==2.8.2
pytz>=2023.3

# JSON handling
orjson==3.9.10

# Mathematical calculations
sympy>=1.12
statsmodels>=0.14.0

# Backtesting and analysis
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0

# API rate limiting
ratelimit==2.2.1
backoff==2.2.1

# Async utilities
asyncio-throttle==1.0.2

# Caching
diskcache==5.6.3

# Environment and system
psutil>=5.9.6
memory-profiler>=0.61.0

# Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development tools (optional)
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1